backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
numpy==1.24.3
pymongo==4.6.2
python-dotenv==1.0.1
gunicorn==21.2.0 